requirements.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
Django==4.0.1
Flask==2.0.2
grpcio==1.43.0
grpcio-tools==1.43.0
idna==3.3
image==1.5.33
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
path==16.3.0
Pillow==9.0.0
protobuf==3.19.3
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
qrcode==7.3.1
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.8
websockets==10.1
Werkzeug==2.0.2
